requirements.txt
streamlit
pandas
python-dotenv
xmlrpc
folium
streamlit-folium
plotly
openpyxl